Resolve #184 Allow User to Submit Multiple Questions
In order to allow a user to create a response to a each question we must create a generated attribute for the title that is a combination of the name and question being answered.

This commit also removes the unused question3 enum.

We also remove the unused story_params method in the events_controller since it should not have been introduced in a previous commit.