Fix Responsive Style Issues
This commit addresses some responsive styling issues by removing the 50% width of .event-start and .event-city so that we do not get weird stacking of the event information. We also add a small margin to the right side of these classes so that we do not have text collisions at certain responsive widths.