Serializer and AJAX for Announcements
This commit updates the serializers in refinery and the carousel javascript in order to use the JSON API standard that we use for most of our other endpoints. Furthermore the responsibility for populating the carousel in the DOM is moved to the Javascript file and it can now dynamically change the items in the carousel based on the clicked number. It will populate a number based on the number of carousel items in the backend.