Adjust initial cards for events
Because filtering by Events displays the “Join us for an event!” card as well, displaying the first 9 cards caused the 9th card to sit alone in a 4th row. Added a conditional to check if the content type is Events and, if so, follows conditional logic to only load the 8 first events.

Fix errors caught in code review