Resolve IFRAME Not Working in Reports 🐛
In 2018 some folks had strong opinions about the appearance of YouTube video embeds on MetroCommon. The result of these feelings lead to the writing of this JavaScript, which surely seemed like a great idea at the time. The script was tested. The embed looked different. Special. Better. More MetroCommony. The script was tested. It worked. Except for a minor bug in MSIE that I briefly fixed. It was merged and we all moved on with our lives.

Fast forward to 2020. The plan is progressing. Content is being built. New kinds of content. Content that doesn’t quite fit into our strange editor, so we call upon our trusty friend the IFRAME! When editors tried to add IFRAMEs they disappeared. Was it a rapture? The Russians? A ghost?

No it was this JavaScript. Sucking away IFRAMEs like a DOM vampire. This commit banishes the vampire from our code base, leaving us to potentially recreate this feature if we ever want it by digging it back out from our git history. Until then we apply the remove dead code refactor here, and solve all our problems.