Fix ul bullets missing in content blocks
This commit fixes the bullets missing in content blocks by migrating the list-style rule into a class. We then apply this class to the relevant footer unordered lists to maintain the aesthetic we know and love, but give content authors the freedom to use bulleted lists.

This commit also applies HTMLPrettify to the _footer.html.erb file to make the formatting consistent.