Resolve #468 Adjust Report Width
Adjust the header image for reports such that on mobile it fills the entire top of the screen. On non-mobile we now occupy 50% of the header space and cease to push the image down below the header material. The act of setting an aspect ratio that lines up with the heading material and does not look overly stretchy on mobile is left as an exercise to the user.

We also adjust the paragraphs so they have spacing between them like you would expect.