Remove Menu Code
Discovered while attempting to refactor jQuery as an import into find-out.js, this code steals responsibility for creating the sidebar menu from the Refinery admin interface and appends these items to the sidebar regardless of whether a user wants it. Removing this code restores agency to the CMS where it belongs.