Resolve #469 Social Cards for Publications
This commit adds custom OpenGraph properties for report publications as an article type. This includes an image based on the OneBox image. I also factored the jQuery and jQuery-nice-select items into the find-out page (aka taggings) since they are only used for that page (as far as I can tell).