Allow User to Add News Link
This commit adds functionality so that a user can add the link to the database for the announcements aka “news” functionality.

It also refactors the responsbility for confirming that an announcement item has tags back into the server.

There are also design changes to the tag interface to make it more clear to the user that the tags should be selected, and provide them with the capability to select those tags.